requirements.txt
openai
python-dotenv
pyttsx3
SpeechRecognition
PyAudio
dateparser